ITS ORGANIZED
see, told you. Got rid of one function to make it more modular and cleaner code. Could be even more modular, but this will not be needed untill i will implement a search function on the website. But this is for later, now we make the front-end for the individual article